Move Sentry bootstrap to separate file
